+++ ThirdPartyAdapters/chartboost/chartboost/src/test/kotlin/com/google/ads/mediation/chartboost/ChartboostAdapterTest.kt
@@ -1,7 +1,7 @@
 package com.google.ads.mediation.chartboost
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
+import kotlin.test.assertIs
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -15,6 +15,6 @@ class ChartboostAdapterTest {
   fun instanceOfChartboostMediationAdapter() {
     adapter = ChartboostAdapter()
 
-    assertThat(adapter is ChartboostMediationAdapter).isTrue()
+    assertIs<ChartboostMediationAdapter>(adapter)
   }
 }