+++ ThirdPartyAdapters/ironsource/ironsource/src/test/kotlin/com/google/ads/mediation/ironsource/IronSourceAdapterTest.kt
@@ -1,7 +1,7 @@
 package com.google.ads.mediation.ironsource
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.common.truth.Truth.assertThat
+import kotlin.test.assertIs
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -15,6 +15,6 @@ class IronSourceAdapterTest {
   fun instanceOfIronSourceAdapter_returnsAnInstanceOfIronSourceMediationAdapter() {
     adapter = IronSourceAdapter()
 
-    assertThat(adapter is IronSourceMediationAdapter).isTrue()
+    assertIs<IronSourceMediationAdapter>(adapter)
   }
 }