+++ ThirdPartyAdapters/meta/meta/src/main/java/com/google/ads/mediation/facebook/rtb/FacebookRtbBannerAd.java
@@ -59,7 +59,7 @@ public class FacebookRtbBannerAd implements MediationBannerAd, AdListener {
 
   public void render(@NonNull MediationBannerAdConfiguration adConfiguration) {
     Bundle serverParameters = adConfiguration.getServerParameters();
-    String placementID = FacebookMediationAdapter.getPlacementID(serverParameters);
+    String placementID = FacebookMediationAdapter.getPlacementId(serverParameters);
     if (TextUtils.isEmpty(placementID)) {
       AdError error = new AdError(ERROR_INVALID_SERVER_PARAMETERS,
           "Failed to request ad. PlacementID is null or empty.", ERROR_DOMAIN);