+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,10 +1,19 @@
 ## InMobi Android Mediation Adapter Changelog
 
+#### Version 11.3.0.0
+- Verified compatibility with InMobi Kotlin SDK version 11.3.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.2.0.
+- Google Mobile Ads Next-Gen SDK version 1.1.0.
+- InMobi Kotlin SDK version 11.3.0.
+
 #### Version 11.2.0.0
 - Verified compatibility with InMobi Kotlin SDK version 11.2.0.
 
 Built and tested with:
 - Google Mobile Ads SDK version 25.1.0.
+- Google Mobile Ads Next-Gen SDK version 0.25.0-beta01.
 - InMobi Kotlin SDK version 11.2.0.
 
 #### Version 11.1.1.1