+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,5 +1,20 @@
 ## IronSource Android Mediation Adapter Changelog
 
+#### Version 9.4.2.0
+- Verified compatibility with ironSource SDK version 9.4.2.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.2.0.
+- Google Mobile Ads Next-Gen SDK version 1.0.1.
+- IronSource SDK version 9.4.2.
+
+#### Version 9.4.1.0
+- Verified compatibility with ironSource SDK version 9.4.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.2.0.
+- IronSource SDK version 9.4.1.
+
 #### Version 9.4.0.0
 - Verified compatibility with ironSource SDK version 9.4.0.