+++ ThirdPartyAdapters/verve/CHANGELOG.md
@@ -1,7 +1,20 @@
 ## Verve Android Mediation Adapter Changelog
 
-#### Version 3.8.1.0 (In progress)
+#### Version 3.8.1.1
+- Filter clickable views to differentiate content info click from regular ad click for VerveNativeAd.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.2.0.
+- Google Mobile Ads Next-Gen SDK version 1.0.1.
+- Verve HyBid SDK version 3.8.1.
+
+#### Version 3.8.1.0
 - Added property to build the adapter with GMA Next-Gen SDK dependency.
+- Verified compatibility with Verve HyBid SDK version 3.8.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 25.1.0.
+- Verve HyBid SDK version 3.8.1.
 
 #### Version 3.7.1.1
 - Removed banner size check from the Verve Adapter to resolve signal collection